fix: change tests location for AVFoundationPlaybackTests, AVFoundationPlaybackAudioTests and AVFoundationPlaybackSubtitleTests
